Un metodo agregado a MateriaData y 3 en InscripcionData
Uno de los metodos de InscripcionData no anda bien :( tambien cambie el nombre de la variable id a idAlumno a ver si resolvia el problema pero no.

+++ UniversidadGrupo36/src/universidadgrupo36/Entidades/Inscripcion.java
@@ -7,18 +7,18 @@ public class Inscripcion {
     private int idInscripcion;
     Alumno alumno;
     Materia materia;
-    double nota;
+    int nota;
 
     public Inscripcion () {}
     
-    public Inscripcion(int idInscripcion, Alumno alumno, Materia materia, double nota) {
+    public Inscripcion(int idInscripcion, Alumno alumno, Materia materia, int nota) {
         this.idInscripcion = idInscripcion;
         this.alumno = alumno;
         this.materia = materia;
         this.nota = nota;
     }
 
-    public Inscripcion(Alumno alumno, Materia materia, double nota) {
+    public Inscripcion(Alumno alumno, Materia materia, int nota) {
         this.alumno = alumno;
         this.materia = materia;
         this.nota = nota;
@@ -52,7 +52,7 @@ public class Inscripcion {
         return nota;
     }
 
-    public void setNota(double nota) {
+    public void setNota(int nota) {
         this.nota = nota;
     }
